+++ fabric/src/main/java/io/github/sakurawald/fuji/module/initializer/skin/structure/SkinSyncer.java
@@ -16,6 +16,7 @@ import net.minecraft.network.packet.s2c.play.ExperienceBarUpdateS2CPacket;
 import net.minecraft.network.packet.s2c.play.PlayerListS2CPacket;
 import net.minecraft.network.packet.s2c.play.PlayerRemoveS2CPacket;
 import net.minecraft.network.packet.s2c.play.UpdateSelectedSlotS2CPacket;
+import net.minecraft.server.PlayerManager;
 import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.server.world.ServerWorld;
 import org.jetbrains.annotations.NotNull;
@@ -63,27 +64,14 @@ public class SkinSyncer {
         /* Restore the previous difficulty. */
         player.networkHandler.sendPacket(new DifficultyS2CPacket(EntityHelper.getServerWorld(player).getDifficulty(), EntityHelper.getServerWorld(player).getLevelProperties().isDifficultyLocked()));
 
-        /* Restore the previous selected slot. */
-        #if MC_VER < MC_1_21_5
-        player.networkHandler.sendPacket(new UpdateSelectedSlotS2CPacket(player.getInventory().selectedSlot));
-        #elif MC_VER >= MC_1_21_5
-        player.networkHandler.sendPacket(new UpdateSelectedSlotS2CPacket(player.getInventory().getSelectedSlot()));
-        #endif
-
-        /* Restore the previous inventory slots. */
-        player.playerScreenHandler.updateToClient();
+        /* Restore the previous inventory slots + selected slot. */
+        ServerHelper.getPlayerManager().sendPlayerStatus(player);
 
         /* Restore the previous abilities. */
         player.sendAbilitiesUpdate();
 
         /* Restore the previous status effects. */
-        for (StatusEffectInstance effect : player.getStatusEffects()) {
-        #if MC_VER <= MC_1_20_4
-        player.networkHandler.sendPacket(new EntityStatusEffectS2CPacket(player.getId(), effect));
-        #elif MC_VER > MC_1_20_4
-        player.networkHandler.sendPacket(new EntityStatusEffectS2CPacket(player.getId(), effect, false));
-        #endif
-        }
+        ServerHelper.getPlayerManager().sendStatusEffects(player);
 
         /* Restore the previous experience. */
         player.networkHandler.sendPacket(new ExperienceBarUpdateS2CPacket(player.experienceProgress, player.totalExperience, player.experienceLevel));
@@ -94,6 +82,9 @@ public class SkinSyncer {
             player.networkHandler.sendPacket(new EntityPassengersSetS2CPacket(vehicle));
         }
         player.networkHandler.sendPacket(new EntityPassengersSetS2CPacket(player));
+
+        /* Send the dimension info to the client to prevent it from getting stuck on the dimension loading screen. */
+        ServerHelper.getPlayerManager().sendWorldInfo(player, PlayerHelper.getServerWorld(player));
 
         /* Update the entity tracker. (Does not harm) */
         player.networkHandler.sendPacket(new EntityTrackerUpdateS2CPacket(player.getId(), player.getDataTracker().getChangedEntries()));